Working program
