+++ app/src/main/java/ru/yandex/practicum/contacts/presentation/base/BaseListDiffCallback.java
@@ -4,6 +4,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.DiffUtil;
 
 public class BaseListDiffCallback <A extends ListDiffInterface<A>> extends DiffUtil.ItemCallback<A> implements ListDiffInterface<A>{
+
     @Override
     public boolean areItemsTheSame(@NonNull A a, @NonNull A t1) {
         return false;
