+++ app/src/main/java/ru/yandex/practicum/contacts/presentation/main/ContactUi.java
@@ -43,6 +43,11 @@ public class ContactUi implements ListDiffInterface<ContactUi> {
     }
 
     @Override
+    public boolean theSameAs(ContactUi contactUi) {
+        return this.hashCode() == contactUi.hashCode();
+    }
+
+    @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
