+++ app/src/main/java/ru/yandex/practicum/contacts/presentation/sort/SortTypeUI.java
@@ -24,6 +24,11 @@ public class SortTypeUI implements ListDiffInterface<SortTypeUI> {
     }
 
     @Override
+    public boolean theSameAs(SortTypeUI sortTypeUI) {
+        return this.getSortType() == sortTypeUI.getSortType();
+    }
+
+    @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
